Import auto backups from new location.
